# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YP API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Log into API",
                "description": "Returns token with User Infos",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of the user",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of the user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Register into API",
                "description": "Returns token with User Infos",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of the user",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of the user",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the user",
                        "required": true
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "Lastname of the user",
                        "required": true
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "description": "Age of the user",
                        "required": true
                    },
                    {
                        "name": "phoneNumber",
                        "in": "query",
                        "description": "Phone Number of the user",
                        "required": true
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Gender of the user",
                        "required": true
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City of the user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}